requirements.txt
langchain
langchain-community
PyPDF2
python-docx